Harmonic (Sinusoidal) Illumination Pattern

SIMIlluminationPatterns.Harmonic — Type

Harmonic (sinusoidal) illumination pattern in the form

\[ I(\vec{r})=1+{\frac{m}{2}}\cos\left(2π⋅ (kₓ⋅(\vec{r})ₓ + k_y ⋅ (\vec{r})_y) + \phi\right)\]

Harmonic(m::Real, θ::Real, ν::Frequency, ϕ::Real)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Frequency,Frequency}, ϕ::Real)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Real,Real}, ϕ::Real, Δxy::Length)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Real,Real}, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Harmonic(m::Real, θ::Real, λ::Length, ϕ::Real)
Harmonic(m::Real, (λ_x, λ_y)::Tuple{Length,Length}, ϕ::Real)
Harmonic(m::Real, θ::Real, λ::Real, ϕ::Real, Δxy::Length)
Harmonic(m::Real, θ::Real, λ::Real, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Harmonic(m::Real, (δ_x, δ_y)::Tuple{Real, Real}, size::Union{Tuple{Real, Real}, Real}, ϕ::Real, Δxy::Length)
Harmonic(m::Real, (δ_x, δ_y)::Tuple{Real, Real}, size::Union{Tuple{Real, Real}, Real}, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Parameters have types Real, Frequency or Length and denote:

m: modulation factor

θ: orientation angle (from the $x$-axis) (\theta)

ν: frequency (\nu)

(ν_x, ν_y): wave vector ('shift' in physical units) ($(ν_x, ν_y) = (\sin(θ) ⋅ ν , \, \cos(θ) ⋅ ν)$)

(δ_x, δ_y): shift in pixels

size: size of acquisition image, from which the shift (δ_x, δ_y) is determined

λ: wavelength (\lambda)

(λ_x, λ_y): wavepeak vector ($(λ_x, λ_y) = (\sin(θ) ⋅ λ , \, \cos(θ) ⋅ λ)$)

ϕ: phase offset (\phi)

Δxy or (Δx, Δy): $x$-axis and $y$-axis pixel sizes

SIMIlluminationPatterns.Harmonic2D — Type

Harmonic2D === Harmonic{2}
Harmonic{2} <: IlluminationPattern{2}

<!– TODO: Add plots explaining the parameters <12-10-23> –>

SIMIlluminationPatterns.Harmonic3D — Type

Harmonic3D === Harmonic{3}
Harmonic{3} <: IlluminationPattern{3}