SIMIlluminationPatterns.Harmonic — Type

Harmonic (sinusoidal) illumination pattern in the form

\[ I(\vec{r})=1+{\frac{m}{2}}\cos\left(2π⋅ (kₓ⋅(\vec{r})ₓ + k_y ⋅ (\vec{r})_y) + \phi\right)\]

Harmonic(m::Real, θ::Real, ν::Frequency, ϕ::Real)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Frequency,Frequency}, ϕ::Real)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Real,Real}, ϕ::Real, Δxy::Length)
Harmonic(m::Real, (ν_x, ν_y)::Tuple{Real,Real}, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Harmonic(m::Real, θ::Real, λ::Length, ϕ::Real)
Harmonic(m::Real, (λ_x, λ_y)::Tuple{Length,Length}, ϕ::Real)
Harmonic(m::Real, θ::Real, λ::Real, ϕ::Real, Δxy::Length)
Harmonic(m::Real, θ::Real, λ::Real, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Harmonic(m::Real, (δ_x, δ_y)::Tuple{Real, Real}, size::Union{Tuple{Real, Real}, Real}, ϕ::Real, Δxy::Length)
Harmonic(m::Real, (δ_x, δ_y)::Tuple{Real, Real}, size::Union{Tuple{Real, Real}, Real}, ϕ::Real, (Δx, Δy)::Tuple{Length,Length})

Parameters have types Real, Frequency or Length and denote:

m: modulation factor

θ: orientation angle (from the $x$-axis) (\theta)

ν: frequency (\nu)

(ν_x, ν_y): wave vector ('shift' in physical units) ($(ν_x, ν_y) = (\sin(θ) ⋅ ν , \, \cos(θ) ⋅ ν)$)

(δ_x, δ_y): shift in pixels

size: size of acquisition image, from which the shift (δ_x, δ_y) is determined

λ: wavelength (\lambda)

(λ_x, λ_y): wavepeak vector ($(λ_x, λ_y) = (\sin(θ) ⋅ λ , \, \cos(θ) ⋅ λ)$)

ϕ: phase offset (\phi)

Δxy or (Δx, Δy): $x$-axis and $y$-axis pixel sizes

SIMIlluminationPatterns.Harmonic2D — Type

Harmonic2D === Harmonic{2}
Harmonic{2} <: IlluminationPattern{2}

SIMIlluminationPatterns.Harmonic3D — Type

Harmonic3D === Harmonic{3}
Harmonic{3} <: IlluminationPattern{3}

SIMIlluminationPatterns.IP — Method

(ip::IP{N})(T::Type{<:Real}; Δxy)::IPR{T,N}
(ip::IP{N})(; Δxy)::IPR{T,N}

Create an illumination pattern realization with pixel dimensions Δxy.

Parameters

Δxy: can be a NTuple{Length} or a Length

julia> h = Harmonic(1.0, π / 4, 2 / 61u"nm", 0.0)
Harmonic2D(m=1.0, θ=0.7853981633974483, ν=0.03278688524590164 nm^-1, ϕ=0.0)

julia> h(;Δxy=30.5u"nm")
Harmonic2D(m=1.0, θ=0.7853981633974483, ν=0.03278688524590164 nm^-1, ϕ=0.0)(Δxy = 30.5 nm) with eltype Float64

julia> h(Float16;Δxy=(33.5u"nm", 30.5u"nm"))
Harmonic2D(m=1.0, θ=0.7853981633974483, ν=0.03278688524590164 nm^-1, ϕ=0.0)(Δxy = (33.5 nm, 30.5 nm)) with eltype Float16

SIMIlluminationPatterns.IlluminationPattern — Type

Super type for N-dimensional illumation patterns. Any particular illumination pattern is a subtype of this. See the documentation for IlluminationPattern interface to see how to implement your own illumination pattern subtype.

See also: Harmonic2D

Implementation

When implementing a new subtype A <: IlluminationPattern{N} you need to implement the following methods:

(pattern::A)(r::Vararg{Length, N})::Real representing the intensity of pattern at the position r

and these optional methods:

TODO

The inner constructor of the subtype A should check that N::Integer unless your subtype has a specified dimension.

struct A1{N} <: IlluminationPattern{N}
    function A{N}() where {N}
        @assert N isa Integer
        new{N}()
    end
end

struct A2 <: IlluminationPattern{2}
    function A2()
        # No need for any assertions
        new()
    end
end

SIMIlluminationPatterns.IlluminationPatternRealization — Type

N-dimensional illumination pattern realization. You can fix the pixels sizes (Δxy) and sample the pattern on your sensor and optical system setup.